[Refactor]James - Refactored and modularized the default home screen
