+++ aislequeue/lib/app.dart
@@ -1,18 +1,14 @@
 import 'package:flutter/material.dart';
 import 'screens/home_screen.dart';
 
-class LayoutCreator extends StatelessWidget {
-  const LayoutCreator({super.key});
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'AisleQueue',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.green),
-        useMaterial3: true,
-      ),
-      home: const MyHomePage(title: "AisleQueue"),
+      debugShowCheckedModeBanner: false,
+      home: HomeScreen(),
     );
   }
 }
