+++ aislequeue/lib/screens/home_screen.dart
@@ -1,426 +1,249 @@
 import 'package:flutter/material.dart';
-import '../widgets/custom_app_bar.dart';
-import '../widgets/grid_view_widget.dart';
-import '../widgets/search_bar_widget.dart' as custom_widgets;
-import '../models/placed_tile_data.dart';
-import '../utils/device_type.dart';
+import '../widgets/aisle_box.dart';
+import '../widgets/bottom_box.dart';
+import '../widgets/center_aisle.dart';
+import '../widgets/circle_num.dart';
+import '../widgets/left_box.dart';
+import '../widgets/right_box.dart';
+import '../utils/app_colors.dart';
 
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  late String deviceType;
-  static const int gridColumns = 32;
-  static const int gridRows = 32;
-  static const double gridCellSize = 40;
-
-  final List<PlacedTileData> _placedTiles = [];
-  bool _isPlacementMode = false;
-  bool _isSelectingFirstPoint = false;
-  bool _isSelectingSecondPoint = false;
-  bool _isRemoveMode = false;
-  int? _firstGridX;
-  int? _firstGridY;
-  int _currentGridX = 0;
-  int _currentGridY = 0;
-  double _currentScale = 1.0;
-
-  final TransformationController _transformationController =
-      TransformationController();
-  final TextEditingController _searchController = TextEditingController();
-  bool _isSearching = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _transformationController.addListener(_onTransformationChanged);
-  }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    deviceType = getDeviceType(context);
-  }
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    _transformationController.removeListener(_onTransformationChanged);
-    _transformationController.dispose();
-    super.dispose();
-  }
-
-  void _onTransformationChanged() {
-    Matrix4 matrix = _transformationController.value;
-    double scaleX = matrix.getColumn(0).xyz.length;
-    double scaleY = matrix.getColumn(1).xyz.length;
-    double newScale = (scaleX + scaleY) / 2;
-
-    if ((newScale - _currentScale).abs() > 0.01) {
-      setState(() {
-        _currentScale = newScale;
-      });
-    }
-  }
-
-  void _filterTiles(String query) {
-    setState(() {
-      _isSearching = query.isNotEmpty;
-    });
-  }
-
-  void _togglePlacementMode() {
-    setState(() {
-      _isPlacementMode = !_isPlacementMode;
-      _isRemoveMode = false;
-      _resetSelection();
-    });
-  }
-
-  void _toggleRemoveMode() {
-    setState(() {
-      _isRemoveMode = !_isRemoveMode;
-      _isPlacementMode = false;
-      _resetSelection();
-    });
-  }
-
-  void _updateGridPosition(Offset position) {
-    if (_isPlacementMode || _isRemoveMode) {
-      final Offset localPosition = _transformationController.toScene(position);
-      setState(() {
-        _currentGridX = (localPosition.dx / gridCellSize).floor();
-        if (deviceType == 'Phone') {
-          _currentGridY =
-              ((localPosition.dy - ((gridCellSize / _currentScale) * 3.50)) /
-                      gridCellSize)
-                  .floor();
-        } else {
-          _currentGridY =
-              ((localPosition.dy - ((gridCellSize / _currentScale) * 3.00)) /
-                      gridCellSize)
-                  .floor();
-        }
-        _currentGridX = _currentGridX.clamp(0, gridColumns - 1);
-        _currentGridY = _currentGridY.clamp(0, gridRows - 1);
-      });
-    }
-  }
-
-  void _handleGridTap() {
-    if (_isPlacementMode) {
-      if (_isSelectingFirstPoint) {
-        setState(() {
-          _firstGridX = _currentGridX;
-          _firstGridY = _currentGridY;
-          _isSelectingFirstPoint = false;
-          _isSelectingSecondPoint = true;
-        });
-      } else if (_isSelectingSecondPoint) {
-        _placeRectangleTiles();
-      }
-    } else if (_isRemoveMode) {
-      _removeOrEditTile();
-    }
-  }
-
-  void _removeOrEditTile() {
-    PlacedTileData? tileToModify = _findTileAtCurrentPosition();
-
-    if (tileToModify != null) {
-      showDialog<void>(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            title: const Text('Tile Options'),
-            content: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                ElevatedButton(
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                    _removeTile(tileToModify);
-                  },
-                  child: const Text('Remove Tile'),
-                ),
-                const SizedBox(height: 10),
-                ElevatedButton(
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                    _showEditCategoryDialog(tileToModify);
-                  },
-                  child: const Text('Edit Category'),
-                ),
-              ],
-            ),
-          );
-        },
-      );
-    }
-  }
-
-  PlacedTileData? _findTileAtCurrentPosition() {
-    try {
-      return _placedTiles.firstWhere(
-        (tile) =>
-            _currentGridX >= tile.gridX &&
-            _currentGridX < tile.gridX + tile.width &&
-            _currentGridY >= tile.gridY &&
-            _currentGridY < tile.gridY + tile.height,
-      );
-    } catch (e) {
-      return null;
-    }
-  }
-
-  void _removeTile(PlacedTileData tile) {
-    setState(() {
-      _placedTiles.remove(tile);
-    });
-  }
-
-  void _showEditCategoryDialog(PlacedTileData tile) {
-    final TextEditingController categoryController =
-        TextEditingController(text: tile.category);
-
-    showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Edit Category'),
-          content: TextField(
-            controller: categoryController,
-            decoration:
-                const InputDecoration(hintText: 'Enter new category name'),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Cancel'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-            TextButton(
-              child: const Text('Save'),
-              onPressed: () {
-                if (categoryController.text.isNotEmpty) {
-                  setState(() {
-                    tile.category = categoryController.text;
-                  });
-                  Navigator.of(context).pop();
-                }
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void _placeRectangleTiles() {
-    if (_firstGridX == null || _firstGridY == null) return;
-
-    int startX = _firstGridX!;
-    int startY = _firstGridY!;
-    int endX = _currentGridX;
-    int endY = _currentGridY;
-
-    if (startX > endX) {
-      int temp = startX;
-      startX = endX;
-      endX = temp;
-    }
-
-    if (startY > endY) {
-      int temp = startY;
-      startY = endY;
-      endY = temp;
-    }
-
-    if (_checkTileOverlap(startX, startY, endX, endY)) {
-      _showPlacementErrorDialog();
-      return;
-    }
-
-    _showRectangleCategoryDialog(startX, startY, endX, endY);
-  }
-
-  bool _checkTileOverlap(int startX, int startY, int endX, int endY) {
-    return _placedTiles.any((tile) {
-      bool xOverlap =
-          !(endX < tile.gridX || startX > (tile.gridX + tile.width - 1));
-      bool yOverlap =
-          !(endY < tile.gridY || startY > (tile.gridY + tile.height - 1));
-      return xOverlap && yOverlap;
-    });
-  }
-
-  void _showPlacementErrorDialog() {
-    showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Placement Error'),
-          content: const Text(
-              'The selected area overlaps with existing tiles. Please choose a different area.'),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('OK'),
-              onPressed: () {
-                Navigator.of(context).pop();
-                _resetSelection();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> _showRectangleCategoryDialog(
-      int startX, int startY, int endX, int endY) async {
-    final TextEditingController categoryController = TextEditingController();
-
-    return showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Enter Category Name for Rectangle'),
-          content: TextField(
-            controller: categoryController,
-            decoration:
-                const InputDecoration(hintText: 'Enter a category name'),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Cancel'),
-              onPressed: () {
-                Navigator.of(context).pop();
-                _resetSelection();
-              },
-            ),
-            TextButton(
-              child: const Text('Add'),
-              onPressed: () {
-                if (categoryController.text.isNotEmpty) {
-                  _placedTiles.removeWhere((tile) =>
-                      tile.gridX >= startX &&
-                      tile.gridX <= endX &&
-                      tile.gridY >= startY &&
-                      tile.gridY <= endY);
-
-                  setState(() {
-                    final newTile = PlacedTileData(
-                      gridX: startX,
-                      gridY: startY,
-                      width: endX - startX + 1,
-                      height: endY - startY + 1,
-                      category: categoryController.text,
-                    );
-                    _placedTiles.add(newTile);
-                  });
-
-                  Navigator.of(context).pop();
-                  _resetSelection();
-                }
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void _resetSelection() {
-    setState(() {
-      _isSelectingFirstPoint = _isPlacementMode;
-      _isSelectingSecondPoint = false;
-      _firstGridX = null;
-      _firstGridY = null;
-    });
-  }
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: CustomAppBar(title: widget.title),
-      body: Column(
-        children: [
-          custom_widgets.SearchBar(
-            controller: _searchController,
-            onChanged: _filterTiles,
-          ),
-          Expanded(
-            child: GridViewWidget(
-              placedTiles: _placedTiles,
-              isSearching: _isSearching,
-              searchQuery: _searchController.text,
-              onGridTap: _handleGridTap,
-              onHover: _updateGridPosition,
-              transformationController: _transformationController,
-              isPlacementMode: _isPlacementMode,
-              isRemoveMode: _isRemoveMode,
-              currentGridX: _currentGridX,
-              currentGridY: _currentGridY,
-              isSelectingFirstPoint: _isSelectingFirstPoint,
-              isSelectingSecondPoint: _isSelectingSecondPoint,
-              firstGridX: _firstGridX,
-              firstGridY: _firstGridY,
-            ),
-          ),
-        ],
+      backgroundColor: AppColors.backgroundColor,
+      appBar: AppBar(
+        backgroundColor: AppColors.backgroundColor,
       ),
-      floatingActionButton: Row(
-        mainAxisAlignment: MainAxisAlignment.end,
+      body: Stack(
         children: [
-          Padding(
-            padding: const EdgeInsets.only(right: 10),
-            child: FloatingActionButton(
-              onPressed: _toggleRemoveMode,
-              tooltip: 'Edit/Remove Tiles',
-              backgroundColor:
-                  _isRemoveMode ? Colors.red[700] : Colors.amber[800],
-              heroTag: 'editRemoveButton',
-              child: Icon(
-                _isRemoveMode ? Icons.close : Icons.edit,
-                color: Colors.white,
-                size: 24,
-              ),
-              elevation: _isRemoveMode ? 8 : 4,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-                side: _isRemoveMode
-                    ? BorderSide(color: Colors.red[900]!, width: 2)
-                    : BorderSide.none,
+          // Menu Icon
+          const Positioned(
+            top: 10,
+            left: 20,
+            child: Icon(Icons.menu),
+          ),
+          // Search Bar
+          Positioned(
+            top: 13,
+            left: 104,
+            child: Container(
+              height: 19,
+              width: 210,
+              decoration: BoxDecoration(
+                color: AppColors.searchBarColor,
+                borderRadius: BorderRadius.circular(5),
               ),
             ),
           ),
-          FloatingActionButton(
-            onPressed: _togglePlacementMode,
-            tooltip: 'Add Tiles',
-            backgroundColor:
-                _isPlacementMode ? Colors.red[700] : Colors.green[700],
-            heroTag: 'placementButton',
-            child: Icon(
-              _isPlacementMode ? Icons.close : Icons.add,
-              color: Colors.white,
-              size: 28,
-            ),
-            elevation: _isPlacementMode ? 8 : 4,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(12),
-              side: _isPlacementMode
-                  ? BorderSide(color: Colors.green[900]!, width: 2)
-                  : BorderSide.none,
-            ),
+          // Cart Icon
+          const Positioned(
+            top: 13,
+            right: 20,
+            child: Icon(Icons.shopping_cart),
           ),
+
+          _buildStandaloneComponents(),
+          _buildAisleBoxes(),
+          _buildCenterBoxes(),
+          _buildLeftBoxes(),
+          _buildRightBoxes(),
+          _buildBottomBoxes(),
+          _buildCircleNumbers(),
         ],
       ),
+    );
+  }
+
+  // Helper methods to organize the numerous positioned widgets
+  Widget _buildStandaloneComponents() {
+    return Stack(
+      children: [
+        //32
+        Positioned(
+          top: 65,
+          left: 265,
+          child: Container(
+            height: 30,
+            width: 95,
+            color: Color(0xff5A967A),
+          ),
+        ),
+        //33
+        Positioned(
+          top: 65,
+          left: 150,
+          child: Container(
+            height: 30,
+            width: 85,
+            color: Color(0xff5A967A),
+          ),
+        ),
+        //34
+        Positioned(
+          top: 65,
+          left: 0,
+          child: Container(
+            height: 90,
+            width: 30,
+            color: Color(0xff5A967A),
+          ),
+        ),
+        //34
+        Positioned(
+          top: 65,
+          left: 30,
+          child: Container(
+            height: 30,
+            width: 75,
+            color: Color(0xff5A967A),
+          ),
+        ),
+        //36
+        Positioned(
+          bottom: 65,
+          left: 0,
+          child: Container(
+            height: 155,
+            width: 30,
+            color: Color(0xff5A967A),
+          ),
+        ),
+      ],
+    );
+  }
+
+  // Similar helper methods for other component groups
+  Widget _buildAisleBoxes() {
+    return Stack(
+      children: [
+        // Aisle Boxes 1-10
+        AisleBox(top: 365, left: 55), // Box 1 415
+        AisleBox(top: 340, left: 55), // Box 2
+        AisleBox(top: 365, right: 60), // Box 3
+        AisleBox(top: 340, right: 60), // Box 4
+        AisleBox(top: 425, left: 55), // Box 5
+        AisleBox(top: 400, left: 55), // Box 6 450
+        AisleBox(top: 425, right: 60), // Box 7
+        AisleBox(top: 400, right: 60), // Box 8
+        AisleBox(top: 485, left: 55), // Box 9
+        AisleBox(top: 460, left: 55), // Box 10
+        AisleBox(top: 485, right: 60), // Box 11
+        AisleBox(top: 460, right: 60), // Box 12
+        AisleBox(top: 545, left: 55), // Box 13
+        AisleBox(top: 520, left: 55), // Box 14
+        AisleBox(top: 545, right: 60), // Box 15
+        AisleBox(top: 520, right: 60), // Box 16
+        //Bottom Aisle
+        AisleBox(bottom: 70, left: 50), // Box 17
+        AisleBox(bottom: 95, left: 50), // Box 18
+        AisleBox(bottom: 70, right: 60), // Box 19
+        AisleBox(bottom: 95, right: 60), // Box 20
+        AisleBox(bottom: 15, left: 50), // Box 21
+        AisleBox(bottom: 40, left: 50), // Box 22
+        AisleBox(bottom: 15, right: 60), // Box 23
+        AisleBox(bottom: 40, right: 60), // Box 24
+      ],
+    );
+  }
+
+  Widget _buildCenterBoxes() {
+    return Stack(
+      children: [
+        CenterAisle(top: 230, right: 80), // Box 28 (Center)
+        CenterAisle(top: 150, right: 80), // Box 29 (Center)
+        CenterAisle(top: 230, left: 110), // Box 30 (Center)
+        CenterAisle(top: 150, left: 110), // Box 31 (Center)
+      ],
+    );
+  }
+
+  Widget _buildLeftBoxes() {
+    return Stack(
+      children: [
+        //Left boxes 35
+        LeftBox(top: 195, left: 0), // Box 35.1
+        LeftBox(top: 245, left: 0), // Box 35.2
+        LeftBox(top: 295, left: 0), // Box 35.3
+        LeftBox(top: 345, left: 0), // Box 35.4
+        LeftBox(top: 395, left: 0), // Box 35.5
+      ],
+    );
+  }
+
+  Widget _buildRightBoxes() {
+    return Stack(
+      children: [
+        //Right Box
+        RightBox(bottom: 80, right: 0), // Box 37.1
+        RightBox(bottom: 145, right: 0), // Box 37.2
+        RightBox(bottom: 210, right: 0), // Box 37.3
+        RightBox(bottom: 275, right: 0), // Box 37.4
+      ],
+    );
+  }
+
+  Widget _buildBottomBoxes() {
+    return Stack(
+      children: [
+        //Bottom Box 25
+        BottomBox(bottom: 190, left: 45), // Box 25.1
+        BottomBox(bottom: 145, left: 45), // Box 25.2
+        BottomBox(bottom: 190, left: 125), // Box 25.3
+        BottomBox(bottom: 145, left: 125), // Box 25.4
+        //
+        BottomBox(bottom: 190, right: 50), // Box 27.1
+        BottomBox(bottom: 145, right: 50), // Box 27.2
+        BottomBox(bottom: 190, right: 130), // Box 26.1
+        BottomBox(bottom: 145, right: 130), // Box 26.1
+      ],
+    );
+  }
+
+  Widget _buildCircleNumbers() {
+    return Stack(
+      children: [
+        //Circle
+        CircleNum(top: 365, left: 45, number: '1'),
+        CircleNum(top: 330, left: 170, number: '2'),
+        CircleNum(top: 365, right: 170, number: '3'),
+        CircleNum(top: 330, right: 45, number: '4'),
+        CircleNum(top: 425, left: 45, number: '5'),
+        CircleNum(top: 390, left: 170, number: '6'),
+        CircleNum(top: 425, right: 170, number: '7'),
+        CircleNum(top: 390, right: 45, number: '8'),
+        CircleNum(top: 485, left: 45, number: '9'),
+        CircleNum(top: 450, left: 170, number: '10'),
+        CircleNum(top: 485, right: 170, number: '11'),
+        CircleNum(top: 450, right: 45, number: '12'),
+        CircleNum(top: 545, left: 45, number: '13'),
+        CircleNum(top: 510, left: 170, number: '14'),
+        CircleNum(top: 545, right: 170, number: '15'),
+        CircleNum(top: 510, right: 45, number: '16'),
+        CircleNum(bottom: 63, left: 35, number: '17'),
+        CircleNum(bottom: 100, left: 160, number: '18'),
+        CircleNum(bottom: 100, right: 45, number: '20'),
+        CircleNum(bottom: 63, right: 170, number: '19'),
+        CircleNum(bottom: 10, left: 35, number: '21'),
+        CircleNum(bottom: 38, left: 160, number: '22'),
+        CircleNum(bottom: 10, right: 170, number: '23'),
+        CircleNum(bottom: 38, right: 45, number: '24'),
+        CircleNum(bottom: 170, left: 100, number: '25'),
+        CircleNum(bottom: 145, right: 105, number: '26'),
+        CircleNum(bottom: 190, right: 105, number: '27'),
+        CircleNum(top: 140, right: 60, number: '28'),
+        CircleNum(top: 220, right: 60, number: '29'),
+        CircleNum(top: 220, left: 100, number: '30'),
+        CircleNum(top: 140, left: 100, number: '31'),
+        CircleNum(top: 60, right: 75, number: '32'),
+        CircleNum(top: 60, right: 200, number: '33'),
+        CircleNum(top: 60, left: 90, number: '34'),
+        CircleNum(top: 303, left: 15, number: '35'),
+        CircleNum(bottom: 120, left: 10, number: '36'),
+        CircleNum(bottom: 195, right: 0, number: '37'),
+      ],
     );
   }
 }
