+++ aislequeue/lib/utils/app_colors.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const Color backgroundColor = Color(0xffFDFDF8);
+  static const Color primaryGreen = Color(0xff5A967A);
+  static const Color searchBarColor = Color(0xff9AC9B1);
+  static const Color circleColor = Color(0xff006769);
+  static const Color circleTextColor = Color(0xffFFF1DB);
+}
